Extract `its` support to a module
